requirements.txt
python-telegram-bot==20.7
requests
yt-dlp
groq
googlesearch-python
PyPDF2